requirements.txt
selenium==4.5.0
python-dotenv==1.0.0
urllib3==1.26.7
googlesearch-python
SQLAlchemy 
Flask-SQLAlchemy
numpy



